refactor: перенести методы добавления/удаления лайка в FilmStorage
